requirements.txt
attrs==25.3.0
cattrs==25.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
flatbuffers==25.2.10
h11==0.16.0
idna==3.10
jh2==5.0.9
niquests==3.15.2
numpy==2.3.3
openmeteo_requests==1.7.2
openmeteo_sdk==1.20.1
pandas==2.3.2
platformdirs==4.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
qh3==1.5.4
requests==2.32.5
requests-cache==1.2.1
retry-requests==2.0.0
six==1.17.0
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
urllib3-future==2.13.909
wassima==2.0.1
